Enemy&Depency Updates

+++ Assets/1-Scripts/OnurScripts/ScriptableObjects/CharacterProperties.cs
@@ -5,7 +5,7 @@ public class CharacterProperties : ScriptableObject
 {
     public float WalkSpeed;
     public float AttackArea;
-    public int LightAttackDamage;
-    public int HeavyAttackDamage;
+    public float LightAttackDamage;
+    public float HeavyAttackDamage;
 
 }
